+++ packages/graph-plugins-shared/package.json
@@ -14,6 +14,7 @@
   },
   "dependencies": {
     "@magic/graph-core-infra": "workspace:*",
+    "@magic/graph": "workspace:*",
     "@magic/utils": "workspace:*",
     "mathjs": "catalog:"
   },
