+++ packages/graph/src/core/transaction/setupTransactionSucceeded.ts
@@ -1,4 +1,5 @@
-import { EventHub } from '../../events/createEventHub.ts';
+import { EventHub } from '@magic/graph-core-infra/events/createEventHub';
+
 import { CoreEventMap } from '../events.ts';
 import { CoreControls } from '../types.ts';
 import { propagateTransactionEvents } from './propagateTransactionEvents.ts';
