+++ packages/graph/tsconfig.json
@@ -6,5 +6,9 @@
     "rewriteRelativeImportExtensions": false
   },
   "references": [{ "path": "../utils" }, { "path": "../graph-core-infra" }],
-  "include": ["src/**/*.ts", "src/**/*.d.ts"]
+  "include": [
+    "src/**/*.ts",
+    "src/**/*.d.ts",
+    "../graph-core-infra/src/actions/types.ts"
+  ]
 }
